*Minor changes to satisfy cppcheck and splint. *Fixed a bug that could cause alternate endings to not import. *Improved the "Edit bend strength" dialog to start with the input focus on the bend strength field. *Improved the "Seek to beat/measure" dialog to empty the number field when launched and fixed a minor issue where escape may have to be pressed multiple times to cancel the dialog. *Fixed a bug where tempo changes could cause the chart to run past the end of the chart audio and get removed from the project. *Fixed a bug with RS2 export where linknext status didn't force chordnotes to keep their chord's defined sustain length unless the chordnote also had bend or slide status.

+++ trunk/src/menu/edit.c
@@ -3238,7 +3238,7 @@ void eof_read_clipboard_note(PACKFILE *fp, EOF_EXTENDED_NOTE *temp_note, unsigne
 		return;
 
 	/* read the note */
-	(void) eof_load_song_string_pf(temp_note->name, fp, namelength);	//Read the note's name, up to the specified number of characters
+	(void) eof_load_song_string_pf(temp_note->name, fp, (size_t)namelength);	//Read the note's name, up to the specified number of characters
 	temp_note->type = pack_getc(fp);	//Read the note's difficulty
 	temp_note->note = pack_getc(fp);	//Read the note bitmask value
 	temp_note->beat = pack_igetl(fp);	//Read the beat the note starts in
